requirements.txt
asgiref==3.2.7
Django==3.0.5
django-apptemplates==1.5
django-debug-toolbar==2.2
django-invitations==1.9.3
gunicorn==20.0.4
Markdown==3.2.1
psycopg2-binary==2.8.5
pytz==2019.3
sqlparse==0.3.1
